Create README.md
Welcome to the C Program Repository! This repository contains a collection of C programs developed by Ayushman Patra, a B.Tech CSE student at Parala Maharaja Engineering College (PMEC). The aim of this repository is to provide a comprehensive set of examples and exercises to help students and developers learn and practice C programming.